+++ SDLGameEngine/Vector2D.h
@@ -16,16 +16,16 @@ using namespace std;
 class Vector2D{
 public:
     Vector2D();
-    Vector2D(float x ,float y): m_x(x),m_y(y){};
+    Vector2D(double x ,double y): m_x(x),m_y(y){};
     ~Vector2D();
     
-    float getX() const { return m_x; }
-    float getY() const { return m_y; }
+    double getX() const { return m_x; }
+    double getY() const { return m_y; }
     
-    void setX(float x) { m_x = x; }
-    void setY(float y) { m_y = y; }
+    void setX(double x) { m_x = x; }
+    void setY(double y) { m_y = y; }
     
-    float length();
+    double length();
     
     void normalize();
     
@@ -45,15 +45,15 @@ public:
         return v1;
     }
     
-    Vector2D operator*(float scalar);
-    Vector2D& operator*=(float scalar);
+    Vector2D operator*(double scalar);
+    Vector2D& operator*=(double scalar);
     
-    Vector2D operator/(float scalar);
-    Vector2D& operator/=(float scalar);
+    Vector2D operator/(double scalar);
+    Vector2D& operator/=(double scalar);
     
 private:
-    float m_x;
-    float m_y;
+    double m_x;
+    double m_y;
